corregido error Byte-Short

+++ src/main/java/es/mdef/gesinalog_API/REST/valoracion/ValoracionController.java
@@ -74,7 +74,7 @@ public class ValoracionController {
 
 	    for (Object[] resultado : valoraciones) {
 	    	
-	      	Byte tipoIncidenciaByte = (Byte) resultado[0];
+	    	Short tipoIncidenciaByte = (Short) resultado[0];
 	       	int ordinalTipo = tipoIncidenciaByte.intValue();
 	       	
 			Tipo tipoIncidencia = Tipo.values()[ordinalTipo];
